requirements.txt
streamlit
pandas
pymongo
plotly
python-dateutil